Fix wiping in production
